up: store reducer 분리
