+++ src/store.js
@@ -1,42 +1,6 @@
 import { createStore } from 'redux';
 
-export const initState = {
-  newId: 100,
-  taskTitle: '',
-  tasks: [{
-    id: 1, title: 'haha',
-  }],
-};
-
-function reducer(state = initState, action) {
-  if (action.type === 'changeTaskTitle') {
-    return {
-      ...state,
-      taskTitle: action.payload.taskTitle,
-    };
-  }
-
-  if (action.type === 'addTask') {
-    const { newId, taskTitle, tasks } = state;
-
-    return {
-      ...state,
-      newId: newId + 1,
-      taskTitle: '',
-      tasks: [...tasks, { id: newId, title: taskTitle }],
-    };
-  }
-
-  if (action.type === 'deleteTask') {
-    const { tasks } = state;
-    return {
-      ...state,
-      tasks: tasks.filter(({ id }) => id !== action.payload.id),
-    };
-  }
-
-  return state;
-}
+import reducer from './reducer';
 
 const store = createStore(reducer);
